+++ gulp/tasks/clean.js
@@ -1,0 +1,7 @@
+var gulp    = require('gulp');
+var del     = require('del');
+var config  = require('../config');
+
+gulp.task('clean', function() {
+    return del([config.dest + '/*']);
+});
