+++ gulp/tasks/watch.js
@@ -1,0 +1,7 @@
+var gulp   = require('gulp');
+var config = require('../config');
+
+gulp.task('watch', function() {
+    gulp.watch(config.src + '/scss/**/*.scss', ['styles']);
+    gulp.watch(config.src + '/js/**/*.js', ['scripts']);
+});
