+++ gulp/util/swallowErrors.js
@@ -1,0 +1,6 @@
+var notify = require("gulp-notify");
+
+module.exports = function swallowError (error) {
+    notify(error);
+    this.emit('end');
+}
