+++ package.json
@@ -1,40 +1,41 @@
 {
-    "name": "symfony",
-    "version": "1.0.0",
-    "license": "MIT",
-    "repository": "",
-    "private": true,
-    "dependencies": {
-        "autoprefixer": "^6.1.1",
-        "browserify": "^12.0.1",
-        "cssnano": "^3.3.2",
-        "debowerify": "^1.3.1",
-        "del": "^2.1.0",
-        "domready": "^1.0.8",
-        "editorconfig-tools": "^0.1.1",
-        "gulp": "^3.9.0",
-        "gulp-environments": "^0.1.1",
-        "gulp-favicons": "^2.1.3",
-        "gulp-imagemin": "^2.4.0",
-        "gulp-load-plugins": "^1.1.0",
-        "gulp-postcss": "^6.0.1",
-        "gulp-sass": "^2.1.0",
-        "gulp-size": "^2.0.0",
-        "gulp-sourcemaps": "^1.6.0",
-        "gulp-uglify": "^1.5.1",
-        "imagemin-gifsicle": "^4.2.0",
-        "imagemin-jpegtran": "^4.3.2",
-        "imagemin-optipng": "^4.3.0",
-        "imagemin-pngquant": "^4.2.0",
-        "postcss": "^5.0.12",
-        "postcss-bem-linter": "^2.3.0",
-        "postcss-reporter": "^1.3.0",
-        "postcss-scss": "^0.1.3",
-        "stylelint": "^2.3.4",
-        "vinyl-buffer": "^1.0.0",
-        "vinyl-source-stream": "^1.1.0"
-    },
-    "scripts": {
-        "postinstall": "editorconfig-tools fix ./package.json >> /dev/null"
-    }
+  "name": "symfony",
+  "version": "1.0.0",
+  "license": "MIT",
+  "repository": "",
+  "private": true,
+  "dependencies": {
+    "autoprefixer": "^6.1.1",
+    "browserify": "^12.0.1",
+    "cssnano": "^3.3.2",
+    "debowerify": "^1.3.1",
+    "del": "^2.1.0",
+    "domready": "^1.0.8",
+    "editorconfig-tools": "^0.1.1",
+    "gulp": "^3.9.0",
+    "gulp-environments": "^0.1.1",
+    "gulp-favicons": "^2.1.3",
+    "gulp-imagemin": "^2.4.0",
+    "gulp-load-plugins": "^1.1.0",
+    "gulp-notify": "^2.2.0",
+    "gulp-postcss": "^6.0.1",
+    "gulp-sass": "^2.1.0",
+    "gulp-size": "^2.0.0",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-uglify": "^1.5.1",
+    "imagemin-gifsicle": "^4.2.0",
+    "imagemin-jpegtran": "^4.3.2",
+    "imagemin-optipng": "^4.3.0",
+    "imagemin-pngquant": "^4.2.0",
+    "postcss": "^5.0.12",
+    "postcss-bem-linter": "^2.3.0",
+    "postcss-reporter": "^1.3.0",
+    "postcss-scss": "^0.1.3",
+    "stylelint": "^2.3.4",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0"
+  },
+  "scripts": {
+    "postinstall": "editorconfig-tools fix ./package.json >> /dev/null"
+  }
 }
